refactor: split cli.ts into modular files
Split the monolithic 2105-line cli.ts into focused modules:

- src/colors.ts — terminal color utilities
- src/config.ts — config loading, constants, paths
- src/auth.ts — wallet auth, x402 client
- src/output.ts — output formatting, table, progress bar
- src/http.ts — HTTP request layer with x402 fallback
- src/help.ts — help text for all commands
- src/commands/ — individual command modules:
  - store.ts, recall.ts, list.ts, memory.ts (get/delete/update)
  - search.ts (search/context/extract/ingest/consolidate)
  - relations.ts, status.ts (status/stats/count/suggested/graph)
  - data.ts (export/import/purge), namespace.ts
  - config.ts (init/config), migrate.ts, browse.ts, completions.ts

cli.ts is now a slim 185-line entry point with command routing.
All 146 tests pass. Build output unchanged.

Closes MEM-93
