+++ src/cli.ts
@@ -11,1956 +11,35 @@
  *   2. Fall back to x402 payment if free tier exhausted
  */
 
-import { x402Client } from '@x402/core/client';
-import { x402HTTPClient } from '@x402/core/http';
-import { ExactEvmScheme } from '@x402/evm/exact/client';
-import { toClientEvmSigner } from '@x402/evm';
-import { privateKeyToAccount } from 'viem/accounts';
 import { parseArgs } from './args.js';
-import type { ParsedArgs } from './args.js';
-import yaml from 'js-yaml';
-import * as fs from 'fs';
-import * as path from 'path';
-import * as os from 'os';
-
-import { createRequire } from 'module';
-const _require = createRequire(import.meta.url);
-const { version: VERSION } = _require('../package.json');
-const CONFIG_DIR = path.join(os.homedir(), '.memoclaw');
-const CONFIG_FILE_JSON = path.join(CONFIG_DIR, 'config.json');
-const CONFIG_FILE_YAML = path.join(CONFIG_DIR, 'config');
-
-// Load persisted config early
-function loadPersistedConfig(): { url?: string; privateKey?: string; namespace?: string } {
-  try {
-    if (fs.existsSync(CONFIG_FILE_JSON)) {
-      return JSON.parse(fs.readFileSync(CONFIG_FILE_JSON, 'utf-8'));
-    }
-  } catch {}
-  return {};
-}
-const _persistedConfig = loadPersistedConfig();
-
-const API_URL = process.env.MEMOCLAW_URL || _persistedConfig.url || 'https://api.memoclaw.com';
-const PRIVATE_KEY = (process.env.MEMOCLAW_PRIVATE_KEY || _persistedConfig.privateKey) as `0x${string}`;
-const CONFIG_DIR_PATH = CONFIG_DIR;
-const CONFIG_FILE = CONFIG_FILE_YAML;
-
-// ─── Colors ──────────────────────────────────────────────────────────────────
-
-const NO_COLOR = !!process.env.NO_COLOR || !process.stdout.isTTY;
-
-const c = {
-  reset: NO_COLOR ? '' : '\x1b[0m',
-  bold: NO_COLOR ? '' : '\x1b[1m',
-  dim: NO_COLOR ? '' : '\x1b[2m',
-  red: NO_COLOR ? '' : '\x1b[31m',
-  green: NO_COLOR ? '' : '\x1b[32m',
-  yellow: NO_COLOR ? '' : '\x1b[33m',
-  blue: NO_COLOR ? '' : '\x1b[34m',
-  magenta: NO_COLOR ? '' : '\x1b[35m',
-  cyan: NO_COLOR ? '' : '\x1b[36m',
-  gray: NO_COLOR ? '' : '\x1b[90m',
-};
-
-// ─── Auth ────────────────────────────────────────────────────────────────────
-
-function ensureAuth() {
-  if (!PRIVATE_KEY) {
-    console.error(`${c.red}Error:${c.reset} MEMOCLAW_PRIVATE_KEY environment variable required`);
-    console.error(`${c.dim}Set it with: export MEMOCLAW_PRIVATE_KEY=0x...${c.reset}`);
-    process.exit(1);
-  }
-}
-
-let _account: ReturnType<typeof privateKeyToAccount> | null = null;
-function getAccount() {
-  if (!_account) {
-    ensureAuth();
-    _account = privateKeyToAccount(PRIVATE_KEY);
-  }
-  return _account;
-}
-
-const account = new Proxy({} as ReturnType<typeof privateKeyToAccount>, {
-  get(_, prop) { return (getAccount() as any)[prop]; }
-});
-
-let _x402Client: x402HTTPClient | null = null;
-function getX402Client() {
-  if (!_x402Client) {
-    const signer = toClientEvmSigner(account);
-    const coreClient = new x402Client()
-      .register('eip155:*', new ExactEvmScheme(signer));
-    _x402Client = new x402HTTPClient(coreClient);
-  }
-  return _x402Client;
-}
-
-async function getWalletAuthHeader(): Promise<string> {
-  const timestamp = Math.floor(Date.now() / 1000);
-  const message = `memoclaw-auth:${timestamp}`;
-  const signature = await account.signMessage({ message });
-  return `${account.address}:${timestamp}:${signature}`;
-}
-
-// ─── Stdin Helper ────────────────────────────────────────────────────────────
-
-async function readStdin(): Promise<string | null> {
-  if (process.stdin.isTTY) return null;
-  const chunks: string[] = [];
-  for await (const chunk of process.stdin) chunks.push(chunk.toString());
-  const text = chunks.join('').trim();
-  return text || null;
-}
-
-// ─── Output Helpers ──────────────────────────────────────────────────────────
-
-/** Global output mode from parsed args */
-let outputJson = false;
-let outputQuiet = false;
-let outputPretty = false;
-let outputFormat: 'json' | 'table' | 'csv' | 'yaml' = 'table';
-let outputTruncate = 0; // 0 = no truncation
-let outputFile: string | null = null; // File path for output
-let noTruncate = false; // --no-truncate flag
-
-/** Write to file or stdout */
-function outputWrite(...parts: string[]) {
-  const line = parts.join(' ');
-  if (outputFile) {
-    fs.appendFileSync(outputFile, line + '\n');
-  } else {
-    console.log(line);
-  }
-}
-
-/** Write error to file or stderr */
-function outputError(...parts: string[]) {
-  const line = parts.join(' ');
-  if (outputFile) {
-    fs.appendFileSync(outputFile, line + '\n');
-  } else {
-    console.error(line);
-  }
-}
-
-/** Global field extraction flag */
-let outputField: string | null = null;
-
-/** Extract a field from data using dot notation (e.g. "memory.content") */
-function extractField(data: any, field: string): any {
-  const parts = field.split('.');
-  let val = data;
-  for (const p of parts) {
-    if (val == null) return undefined;
-    val = val[p];
-  }
-  return val;
-}
-
-function out(data: any) {
-  if (outputQuiet) return;
-  // --field extraction: pull a single field and print it
-  if (outputField) {
-    const val = extractField(data, outputField);
-    if (val === undefined) return;
-    if (typeof val === 'object') {
-      outputWrite(JSON.stringify(val, outputPretty ? null : undefined, outputPretty ? 2 : undefined));
-    } else {
-      outputWrite(String(val));
-    }
-    return;
-  }
-  if (outputJson || outputFormat === 'json') {
-    outputWrite(JSON.stringify(data, outputPretty ? null : undefined, outputPretty ? 2 : undefined));
-  } else if (outputFormat === 'yaml') {
-    outputWrite(yaml.dump(data, { indent: 2, lineWidth: 120 }));
-  } else if (outputFormat === 'csv') {
-    if (Array.isArray(data)) {
-      if (data.length === 0) return;
-      const headers = Object.keys(data[0]);
-      outputWrite(headers.join(','));
-      for (const row of data) {
-        outputWrite(headers.map(h => {
-          const val = row[h];
-          const str = val === null || val === undefined ? '' : String(val);
-          return str.includes(',') || str.includes('"') ? `"${str.replace(/"/g, '""')}"` : str;
-        }).join(','));
-      }
-    } else if (typeof data === 'string') {
-      outputWrite(data);
-    } else {
-      outputWrite(JSON.stringify(data, null, 2));
-    }
-  } else if (typeof data === 'string') {
-    outputWrite(data);
-  } else {
-    outputWrite(JSON.stringify(data, null, 2));
-  }
-}
-
-function success(msg: string) {
-  if (outputQuiet) return;
-  if (outputJson) return; // JSON mode suppresses decorative output
-  outputWrite(`${c.green}✓${c.reset} ${msg}`);
-}
-
-function warn(msg: string) {
-  outputError(`${c.yellow}⚠${c.reset} ${msg}`);
-}
-
-function info(msg: string) {
-  if (outputQuiet) return;
-  if (outputJson) return;
-  outputError(`${c.blue}ℹ${c.reset} ${msg}`);
-}
-
-/** Print a simple table */
-function table(rows: Record<string, any>[], columns?: { key: string; label: string; width?: number }[]) {
-  if (rows.length === 0) return;
-  
-  if (outputJson || outputFormat === 'json') {
-    outputWrite(JSON.stringify(rows, null, 2));
-    return;
-  }
-
-  if (outputFormat === 'yaml') {
-    outputWrite(yaml.dump(rows, { indent: 2, lineWidth: 120 }));
-    return;
-  }
-
-  if (outputFormat === 'csv') {
-    out(rows);
-    return;
-  }
-
-  if (!columns) {
-    // Auto-detect columns from first row
-    columns = Object.keys(rows[0]).map(k => ({ key: k, label: k.toUpperCase() }));
-  }
-
-  // Calculate widths
-  const capWidth = args.wide ? 120 : 60;
-  for (const col of columns) {
-    if (!col.width) {
-      col.width = Math.max(
-        col.label.length,
-        ...rows.map(r => String(r[col.key] ?? '').length)
-      );
-      col.width = Math.min(col.width, capWidth); // cap
-    }
-  }
-
-  // Header
-  const header = columns.map(col => col.label.padEnd(col.width!)).join('  ');
-  outputWrite(`${c.bold}${header}${c.reset}`);
-  outputWrite(`${c.dim}${columns.map(col => '─'.repeat(col.width!)).join('──')}${c.reset}`);
-
-  // Rows
-  for (const row of rows) {
-    const line = columns!.map(col => {
-      const val = String(row[col.key] ?? '');
-      return val.length > col.width! ? val.slice(0, col.width! - 1) + '…' : val.padEnd(col.width!);
-    }).join('  ');
-    outputWrite(line);
-  }
-}
-
-/** Simple progress bar */
-function progressBar(current: number, total: number, width = 30): string {
-  const pct = Math.min(current / total, 1);
-  const filled = Math.round(pct * width);
-  return `${c.green}${'█'.repeat(filled)}${c.dim}${'░'.repeat(width - filled)}${c.reset} ${current}/${total}`;
-}
-
-/** Truncate text to specified width */
-function truncate(text: string, width: number): string {
-  if (width <= 0 || text.length <= width) return text;
-  return text.slice(0, width - 1) + '…';
-}
-
-// ─── HTTP ────────────────────────────────────────────────────────────────────
-
-async function request(method: string, path: string, body: any = null) {
-  const url = `${API_URL}${path}`;
-  const headers: Record<string, string> = { 'Content-Type': 'application/json' };
-  const options: RequestInit = { method, headers };
-  if (body) options.body = JSON.stringify(body);
-
-  const walletAuth = await getWalletAuthHeader();
-  headers['x-wallet-auth'] = walletAuth;
-
-  // Apply timeout if configured
-  const controller = new AbortController();
-  const timeoutId = setTimeout(() => controller.abort(), TIMEOUT_MS);
-
-  let res: Response;
-  try {
-    res = await fetch(url, { ...options, headers, signal: controller.signal });
-  } catch (e: any) {
-    clearTimeout(timeoutId);
-    if (e.code === 'ECONNREFUSED' || e.cause?.code === 'ECONNREFUSED') {
-      throw new Error(`Cannot connect to ${API_URL} — is the server running?`);
-    }
-    if (e.code === 'ENOTFOUND' || e.cause?.code === 'ENOTFOUND') {
-      throw new Error(`DNS lookup failed for ${API_URL} — check your internet connection`);
-    }
-    if (e.name === 'AbortError') {
-      throw new Error(`Request timed out after ${TIMEOUT_MS / 1000}s`);
-    }
-    throw new Error(`Network error: ${e.message}`);
-  }
-  clearTimeout(timeoutId);
-
-  const freeTierRemaining = res.headers.get('x-free-tier-remaining');
-  if (freeTierRemaining !== null && process.env.DEBUG) {
-    console.error(`${c.dim}Free tier remaining: ${freeTierRemaining}${c.reset}`);
-  }
-
-  if (res.status === 402) {
-    const errorBody = await res.json();
-    if (process.env.DEBUG) {
-      console.error('=== 402 Response (switching to x402) ===');
-      console.error('Headers:', Object.fromEntries(res.headers.entries()));
-      console.error('Body:', JSON.stringify(errorBody, null, 2));
-    }
-    
-    try {
-      const client = getX402Client();
-      const paymentRequired = client.getPaymentRequiredResponse(
-        (name: string) => res.headers.get(name),
-        errorBody
-      );
-      if (process.env.DEBUG) console.error('Payment required parsed:', JSON.stringify(paymentRequired, null, 2));
-      
-      const paymentPayload = await client.createPaymentPayload(paymentRequired);
-      if (process.env.DEBUG) console.error('Payment payload created');
-      
-      const paymentHeaders = client.encodePaymentSignatureHeader(paymentPayload);
-      if (process.env.DEBUG) console.error('Payment headers:', paymentHeaders);
-      
-      res = await fetch(url, {
-        method,
-        headers: { 'Content-Type': 'application/json', ...paymentHeaders },
-        body: body ? JSON.stringify(body) : undefined,
-      });
-      
-      if (process.env.DEBUG) {
-        console.error('=== Retry Response ===');
-        console.error('Status:', res.status);
-        if (res.status !== 200) {
-          const retryBody = await res.clone().text();
-          console.error('Body:', retryBody);
-        }
-      }
-    } catch (paymentError: any) {
-      if (process.env.DEBUG) console.error('x402 payment failed:', paymentError);
-      throw new Error(
-        `Free tier exhausted. Run \`memoclaw status\` to check usage, or visit memoclaw.com/pricing for paid plans.\n` +
-        `${c.dim}(x402 payment failed: ${paymentError.message})${c.reset}`
-      );
-    }
-  }
-
-  const data = await res.json();
-  if (!res.ok) {
-    throw new Error((data as any).error?.message || `HTTP ${res.status}`);
-  }
-  return data;
-}
-
-// ─── Commands ────────────────────────────────────────────────────────────────
-
-async function cmdStore(content: string, opts: ParsedArgs) {
-  const body: Record<string, any> = { content };
-  if (opts.importance != null && opts.importance !== true) body.importance = parseFloat(opts.importance);
-  if (opts.tags) body.metadata = { tags: opts.tags.split(',').map((t: string) => t.trim()) };
-  if (opts.namespace) body.namespace = opts.namespace;
-
-  const result = await request('POST', '/v1/store', body);
-  if (outputJson) {
-    out(result);
-  } else {
-    success(`Memory stored${result.id ? ` (${c.cyan}${result.id}${c.reset})` : ''}`);
-    if (result.importance !== undefined) info(`Importance: ${result.importance}`);
-  }
-}
-
-async function cmdRecall(query: string, opts: ParsedArgs) {
-  const body: Record<string, any> = { query };
-  if (opts.limit != null && opts.limit !== true) body.limit = parseInt(opts.limit);
-  if (opts.minSimilarity != null && opts.minSimilarity !== true) body.min_similarity = parseFloat(opts.minSimilarity);
-  if (opts.namespace) body.namespace = opts.namespace;
-  if (opts.tags) body.filters = { tags: opts.tags.split(',').map((t: string) => t.trim()) };
-
-  // Watch mode: continuously poll for changes
-  if (opts.watch) {
-    let lastCount = -1;
-    const pollInterval = parseInt(opts.watchInterval || '5000');
-    const readline = await import('readline');
-    const rl = readline.createInterface({ input: process.stdin, output: process.stdout });
-    
-    console.log(`${c.dim}Watching for changes... Press Ctrl+C to stop.${c.reset}`);
-    
-    while (true) {
-      try {
-        const result = await request('POST', '/v1/recall', body) as any;
-        const memories = result.memories || [];
-        
-        // Only output if count changed
-        if (memories.length !== lastCount) {
-          // Clear previous output (simple approach - just print separator)
-          if (lastCount >= 0) {
-            console.log(`${c.dim}${'─'.repeat(40)}${c.reset}`);
-          }
-          lastCount = memories.length;
-          
-          if (memories.length === 0) {
-            console.log(`${c.dim}No memories found.${c.reset}`);
-          } else {
-            for (const mem of memories) {
-              const sim = mem.similarity?.toFixed(3) || '???';
-              const simColor = (mem.similarity || 0) > 0.8 ? c.green : (mem.similarity || 0) > 0.5 ? c.yellow : c.red;
-              const content = outputTruncate ? truncate(mem.content, outputTruncate) : mem.content;
-              console.log(`${simColor}[${sim}]${c.reset} ${content}`);
-              if (mem.metadata?.tags?.length) {
-                console.log(`  ${c.dim}tags: ${mem.metadata.tags.join(', ')}${c.reset}`);
-              }
-            }
-            console.log(`${c.dim}─ ${memories.length} result${memories.length !== 1 ? 's' : ''}${c.reset}`);
-          }
-        }
-        
-        // Wait before next poll
-        await new Promise(r => setTimeout(r, pollInterval));
-      } catch (e) {
-        if (process.env.DEBUG) console.error('Watch error:', e);
-        await new Promise(r => setTimeout(r, pollInterval));
-      }
-    }
-    // unreachable in watch mode — while(true) loop above never exits normally
-  }
-  
-  const result = await request('POST', '/v1/recall', body) as any;
-  
-  if (outputJson) {
-    out(result);
-  } else if (opts.raw) {
-    const memories = result.memories || [];
-    for (const mem of memories) {
-      console.log(mem.content);
-    }
-  } else {
-    const memories = result.memories || [];
-    if (memories.length === 0) {
-      console.log(`${c.dim}No memories found.${c.reset}`);
-    } else {
-      for (const mem of memories) {
-        const sim = mem.similarity?.toFixed(3) || '???';
-        const simColor = (mem.similarity || 0) > 0.8 ? c.green : (mem.similarity || 0) > 0.5 ? c.yellow : c.red;
-        const content = outputTruncate ? truncate(mem.content, outputTruncate) : mem.content;
-        console.log(`${simColor}[${sim}]${c.reset} ${content}`);
-        if (mem.metadata?.tags?.length) {
-          console.log(`  ${c.dim}tags: ${mem.metadata.tags.join(', ')}${c.reset}`);
-        }
-        if (mem.id) {
-          console.log(`  ${c.dim}id: ${mem.id}${c.reset}`);
-        }
-      }
-      console.log(`${c.dim}─ ${memories.length} result${memories.length !== 1 ? 's' : ''}${c.reset}`);
-    }
-  }
-}
-
-async function cmdList(opts: ParsedArgs) {
-  const params = new URLSearchParams();
-  if (opts.limit != null && opts.limit !== true) params.set('limit', opts.limit);
-  if (opts.offset != null && opts.offset !== true) params.set('offset', opts.offset);
-  if (opts.namespace) params.set('namespace', opts.namespace);
-
-  // Watch mode: continuously poll for changes
-  if (opts.watch) {
-    let lastTotal = -1;
-    const pollInterval = parseInt(opts.watchInterval || '5000');
-    
-    console.log(`${c.dim}Watching for changes... Press Ctrl+C to stop.${c.reset}`);
-    
-    while (true) {
-      try {
-        const result = await request('GET', `/v1/memories?${params}`) as any;
-        const memories = result.memories || result.data || [];
-        const total = result.total ?? memories.length;
-        
-        // Only output if total changed
-        if (total !== lastTotal) {
-          if (lastTotal >= 0) {
-            console.log(`${c.dim}${'─'.repeat(40)}${c.reset}`);
-          }
-          lastTotal = total;
-          
-          if (memories.length === 0) {
-            console.log(`${c.dim}No memories found.${c.reset}`);
-          } else {
-            const truncateWidth = outputTruncate || 50;
-            const rows = memories.map((m: any) => ({
-              id: m.id?.slice(0, 8) || '?',
-              content: m.content?.length > truncateWidth ? m.content.slice(0, truncateWidth) + '…' : (m.content || ''),
-              importance: m.importance?.toFixed(2) || '-',
-              tags: m.metadata?.tags?.join(', ') || '',
-              created: m.created_at ? new Date(m.created_at).toLocaleDateString() : '',
-            }));
-            table(rows, [
-              { key: 'id', label: 'ID', width: 10 },
-              { key: 'content', label: 'CONTENT', width: outputTruncate || 52 },
-              { key: 'importance', label: 'IMP', width: 5 },
-              { key: 'tags', label: 'TAGS', width: 20 },
-              { key: 'created', label: 'CREATED', width: 12 },
-            ]);
-            console.log(`${c.dim}─ ${memories.length} of ${total} memories${c.reset}`);
-          }
-        }
-        
-        await new Promise(r => setTimeout(r, pollInterval));
-      } catch (e) {
-        if (process.env.DEBUG) console.error('Watch error:', e);
-        await new Promise(r => setTimeout(r, pollInterval));
-      }
-    }
-  }
-  
-  const result = await request('GET', `/v1/memories?${params}`) as any;
-  
-  if (outputJson) {
-    out(result);
-  } else {
-    let memories = result.memories || result.data || [];
-    
-    // Apply client-side sorting if --sort-by specified
-    if (opts.sortBy && memories.length > 0) {
-      const sortKey = opts.sortBy;
-      const reverse = !!opts.reverse;
-      memories = [...memories].sort((a: any, b: any) => {
-        let aVal = a[sortKey];
-        let bVal = b[sortKey];
-        // Handle nested fields like metadata.tags
-        if (aVal === undefined && sortKey.includes('.')) {
-          const parts = sortKey.split('.');
-          let obj: any = a;
-          for (const p of parts) obj = obj?.[p];
-          aVal = obj;
-          obj = b;
-          for (const p of parts) obj = obj?.[p];
-          bVal = obj;
-        }
-        // Handle dates
-        if (aVal?.includes?.('-') && !isNaN(Date.parse(aVal))) {
-          aVal = new Date(aVal).getTime();
-          bVal = new Date(bVal as string).getTime();
-        }
-        // Handle importance as number
-        if (sortKey === 'importance') {
-          aVal = parseFloat(aVal) || 0;
-          bVal = parseFloat(bVal) || 0;
-        }
-        if (aVal < bVal) return reverse ? 1 : -1;
-        if (aVal > bVal) return reverse ? -1 : 1;
-        return 0;
-      });
-    }
-    
-    if (memories.length === 0) {
-      console.log(`${c.dim}No memories found.${c.reset}`);
-    } else {
-      const truncateWidth = outputTruncate || 50;
-      
-      // Default columns
-      let columns = [
-        { key: 'id', label: 'ID', width: 10 },
-        { key: 'content', label: 'CONTENT', width: outputTruncate || 52 },
-        { key: 'importance', label: 'IMP', width: 5 },
-        { key: 'tags', label: 'TAGS', width: 20 },
-        { key: 'created', label: 'CREATED', width: 12 },
-      ];
-      
-      // Column selection: --columns id,tags,importance
-      if (opts.columns) {
-        const selected = opts.columns.split(',').map((c: string) => c.trim());
-        const colMap: Record<string, { key: string; label: string; width?: number }> = {
-          id: { key: 'id', label: 'ID', width: 10 },
-          content: { key: 'content', label: 'CONTENT', width: outputTruncate || 52 },
-          importance: { key: 'importance', label: 'IMP', width: 5 },
-          tags: { key: 'tags', label: 'TAGS', width: 20 },
-          created: { key: 'created', label: 'CREATED', width: 12 },
-          updated: { key: 'updated', label: 'UPDATED', width: 12 },
-          namespace: { key: 'namespace', label: 'NAMESPACE', width: 15 },
-          type: { key: 'memory_type', label: 'TYPE', width: 10 },
-        };
-        columns = selected.map((k: string) => colMap[k] || { key: k, label: k.toUpperCase(), width: 20 });
-      }
-      
-      const rows = memories.map((m: any) => {
-        const row: Record<string, any> = {};
-        for (const col of columns) {
-          let val = m[col.key];
-          if (col.key === 'content' && val?.length > (col.width || 50)) {
-            val = val.slice(0, (col.width || 50) - 1) + '…';
-          } else if (col.key === 'importance' && val !== undefined) {
-            val = val.toFixed(2);
-          } else if (col.key === 'tags') {
-            val = m.metadata?.tags?.join(', ') || '';
-          } else if ((col.key === 'created' || col.key === 'updated') && m[`${col.key}_at`]) {
-            val = new Date(m[`${col.key}_at`]).toLocaleDateString();
-          } else if (val === undefined || val === null) {
-            val = '';
-          } else {
-            val = String(val);
-          }
-          row[col.key] = val;
-        }
-        return row;
-      });
-      table(rows, columns);
-      if (result.total !== undefined) {
-        console.log(`${c.dim}─ ${memories.length} of ${result.total} memories${c.reset}`);
-      }
-    }
-  }
-}
-
-async function cmdGet(id: string) {
-  const result = await request('GET', `/v1/memories/${id}`) as any;
-  if (outputJson) {
-    out(result);
-  } else {
-    const mem = result.memory || result;
-    console.log(`${c.bold}ID:${c.reset}         ${mem.id || id}`);
-    console.log(`${c.bold}Content:${c.reset}    ${mem.content}`);
-    if (mem.importance !== undefined) console.log(`${c.bold}Importance:${c.reset} ${mem.importance}`);
-    if (mem.namespace) console.log(`${c.bold}Namespace:${c.reset}  ${mem.namespace}`);
-    if (mem.metadata?.tags?.length) console.log(`${c.bold}Tags:${c.reset}       ${mem.metadata.tags.join(', ')}`);
-    if (mem.memory_type) console.log(`${c.bold}Type:${c.reset}       ${mem.memory_type}`);
-    if (mem.created_at) console.log(`${c.bold}Created:${c.reset}    ${new Date(mem.created_at).toLocaleString()}`);
-    if (mem.updated_at) console.log(`${c.bold}Updated:${c.reset}    ${new Date(mem.updated_at).toLocaleString()}`);
-    if (mem.pinned) console.log(`${c.bold}Pinned:${c.reset}     ${c.green}yes${c.reset}`);
-  }
-}
-
-async function cmdDelete(id: string) {
-  const result = await request('DELETE', `/v1/memories/${id}`);
-  if (outputJson) {
-    out(result);
-  } else {
-    success(`Memory ${c.cyan}${id.slice(0, 8)}…${c.reset} deleted`);
-  }
-}
-
-async function cmdSuggested(opts: ParsedArgs) {
-  const params = new URLSearchParams();
-  if (opts.limit != null && opts.limit !== true) params.set('limit', opts.limit);
-  if (opts.namespace) params.set('namespace', opts.namespace);
-  if (opts.category) params.set('category', opts.category);
-
-  const result = await request('GET', `/v1/suggested?${params}`) as any;
-  
-  if (outputJson) {
-    out(result);
-  } else {
-    if (result.categories) {
-      const cats = Object.entries(result.categories)
-        .map(([k, v]) => `${c.bold}${k}${c.reset}=${v}`).join('  ');
-      console.log(`Categories: ${cats}`);
-      console.log(`${c.dim}${'─'.repeat(60)}${c.reset}`);
-    }
-    
-    const suggestions = result.suggested || [];
-    if (suggestions.length === 0) {
-      console.log(`${c.dim}No suggested memories.${c.reset}`);
-    } else {
-      for (const mem of suggestions) {
-        const cat = mem.category?.toUpperCase() || '???';
-        const catColor = { STALE: c.red, FRESH: c.green, HOT: c.yellow, DECAYING: c.magenta }[cat] || c.gray;
-        const text = mem.content.length > 100 ? mem.content.slice(0, 100) + '…' : mem.content;
-        console.log(`${catColor}[${cat}]${c.reset} ${c.dim}(${mem.review_score?.toFixed(2) || '?'})${c.reset} ${text}`);
-        if (mem.metadata?.tags?.length) {
-          console.log(`  ${c.dim}tags: ${mem.metadata.tags.join(', ')}${c.reset}`);
-        }
-      }
-    }
-  }
-}
-
-async function cmdUpdate(id: string, opts: ParsedArgs) {
-  const body: Record<string, any> = {};
-  if (opts.content) body.content = opts.content;
-  if (opts.importance != null && opts.importance !== true) body.importance = parseFloat(opts.importance);
-  if (opts.memoryType) body.memory_type = opts.memoryType;
-  if (opts.namespace) body.namespace = opts.namespace;
-  if (opts.tags) body.metadata = { tags: opts.tags.split(',').map((t: string) => t.trim()) };
-  if (opts.expiresAt) body.expires_at = opts.expiresAt;
-  if (opts.pinned !== undefined) body.pinned = opts.pinned === 'true' || opts.pinned === true;
-
-  if (Object.keys(body).length === 0) {
-    throw new Error('No fields to update. Use --content, --importance, --tags, etc.');
-  }
-
-  const result = await request('PATCH', `/v1/memories/${id}`, body);
-  if (outputJson) {
-    out(result);
-  } else {
-    success(`Memory ${c.cyan}${id.slice(0, 8)}…${c.reset} updated`);
-  }
-}
-
-async function cmdIngest(opts: ParsedArgs) {
-  const body: Record<string, any> = {};
-  if (opts.text) body.text = opts.text;
-  if (opts.namespace) body.namespace = opts.namespace;
-  if (opts.sessionId) body.session_id = opts.sessionId;
-  if (opts.agentId) body.agent_id = opts.agentId;
-  if (opts.autoRelate !== undefined) body.auto_relate = opts.autoRelate !== 'false';
-  else body.auto_relate = true;
-
-  if (!body.text) {
-    const stdin = await readStdin();
-    if (stdin) body.text = stdin;
-  }
-
-  if (!body.text) throw new Error('Text required (use --text or pipe via stdin)');
-
-  const result = await request('POST', '/v1/ingest', body) as any;
-  if (outputJson) {
-    out(result);
-  } else {
-    const count = result.memories_created ?? result.count ?? '?';
-    success(`Ingested text → ${count} memories created`);
-  }
-}
-
-async function cmdExtract(text: string, opts: ParsedArgs) {
-  const body: Record<string, any> = { text };
-  if (opts.namespace) body.namespace = opts.namespace;
-  if (opts.sessionId) body.session_id = opts.sessionId;
-  if (opts.agentId) body.agent_id = opts.agentId;
-
-  const result = await request('POST', '/v1/memories/extract', body);
-  out(result);
-}
-
-async function cmdSearch(query: string, opts: ParsedArgs) {
-  const params = new URLSearchParams({ q: query });
-  if (opts.limit != null && opts.limit !== true) params.set('limit', opts.limit);
-  if (opts.namespace) params.set('namespace', opts.namespace);
-  if (opts.tags) params.set('tags', opts.tags);
-
-  const result = await request('GET', `/v1/memories/search?${params}`) as any;
-
-  if (outputJson) {
-    out(result);
-  } else if (opts.raw) {
-    const memories = result.memories || result.data || [];
-    for (const mem of memories) {
-      console.log(mem.content);
-    }
-  } else {
-    const memories = result.memories || result.data || [];
-    if (memories.length === 0) {
-      console.log(`${c.dim}No memories found.${c.reset}`);
-    } else {
-      const truncateWidth = outputTruncate || 80;
-      for (const mem of memories) {
-        const content = noTruncate ? mem.content : truncate(mem.content || '', truncateWidth);
-        console.log(`${c.cyan}${(mem.id || '?').slice(0, 8)}${c.reset}  ${content}`);
-        if (mem.metadata?.tags?.length) {
-          console.log(`  ${c.dim}tags: ${mem.metadata.tags.join(', ')}${c.reset}`);
-        }
-      }
-      console.log(`${c.dim}─ ${memories.length} result${memories.length !== 1 ? 's' : ''} (text search, free)${c.reset}`);
-    }
-  }
-}
-
-async function cmdContext(query: string, opts: ParsedArgs) {
-  const body: Record<string, any> = { query };
-  if (opts.namespace) body.namespace = opts.namespace;
-  if (opts.limit != null && opts.limit !== true) body.limit = parseInt(opts.limit);
-
-  const result = await request('POST', '/v1/context', body) as any;
-
-  if (outputJson) {
-    out(result);
-  } else {
-    const context = result.context || result.text || result.content;
-    if (context) {
-      console.log(context);
-    } else {
-      // Fallback: print the whole result
-      out(result);
-    }
-  }
-}
-
-async function cmdConsolidate(opts: ParsedArgs) {
-  const body: Record<string, any> = {};
-  if (opts.namespace) body.namespace = opts.namespace;
-  if (opts.minSimilarity != null && opts.minSimilarity !== true) body.min_similarity = parseFloat(opts.minSimilarity);
-  if (opts.mode) body.mode = opts.mode;
-  if (opts.dryRun !== undefined) body.dry_run = true;
-
-  const result = await request('POST', '/v1/memories/consolidate', body) as any;
-  if (outputJson) {
-    out(result);
-  } else {
-    if (opts.dryRun) {
-      info('Dry run — no changes applied');
-    }
-    const merged = result.merged_count ?? result.merged ?? '?';
-    success(`Consolidated: ${merged} memories merged`);
-    if (result.clusters) {
-      info(`Clusters found: ${result.clusters.length}`);
-    }
-  }
-}
-
-async function cmdRelations(subcmd: string, rest: string[], opts: ParsedArgs) {
-  if (subcmd === 'list') {
-    if (!rest[0]) throw new Error('Memory ID required');
-    const result = await request('GET', `/v1/memories/${rest[0]}/relations`) as any;
-    if (outputJson) {
-      out(result);
-    } else {
-      const relations = result.relations || [];
-      if (relations.length === 0) {
-        console.log(`${c.dim}No relations found.${c.reset}`);
-      } else {
-        const rows = relations.map((r: any) => ({
-          id: r.id?.slice(0, 8) || '?',
-          type: r.relation_type || '?',
-          target: r.target_id?.slice(0, 8) || '?',
-        }));
-        table(rows, [
-          { key: 'id', label: 'ID', width: 10 },
-          { key: 'type', label: 'TYPE', width: 16 },
-          { key: 'target', label: 'TARGET', width: 10 },
-        ]);
-      }
-    }
-  } else if (subcmd === 'create') {
-    if (!rest[0] || !rest[1] || !rest[2]) throw new Error('Usage: relations create <memory-id> <target-id> <type>');
-    const validTypes = ['related_to', 'derived_from', 'contradicts', 'supersedes', 'supports'];
-    if (!validTypes.includes(rest[2])) {
-      throw new Error(`Invalid relation type "${rest[2]}". Valid: ${validTypes.join(', ')}`);
-    }
-    const body = { target_id: rest[1], relation_type: rest[2] };
-    const result = await request('POST', `/v1/memories/${rest[0]}/relations`, body);
-    if (outputJson) {
-      out(result);
-    } else {
-      success(`Relation created: ${rest[0].slice(0, 8)}… ${c.cyan}${rest[2]}${c.reset} → ${rest[1].slice(0, 8)}…`);
-    }
-  } else if (subcmd === 'delete') {
-    if (!rest[0] || !rest[1]) throw new Error('Usage: relations delete <memory-id> <relation-id>');
-    const result = await request('DELETE', `/v1/memories/${rest[0]}/relations/${rest[1]}`);
-    if (outputJson) {
-      out(result);
-    } else {
-      success('Relation deleted');
-    }
-  } else {
-    throw new Error('Usage: relations [list|create|delete]');
-  }
-}
-
-async function cmdStatus() {
-  const walletAuth = await getWalletAuthHeader();
-  const res = await fetch(`${API_URL}/v1/free-tier/status`, {
-    headers: { 'x-wallet-auth': walletAuth }
-  });
-  
-  if (res.ok) {
-    const data = await res.json() as any;
-    if (outputJson) {
-      out(data);
-    } else {
-      console.log(`${c.bold}Wallet:${c.reset}     ${data.wallet}`);
-      const remaining = data.free_tier_remaining ?? 0;
-      const total = data.free_tier_total ?? 1000;
-      const pct = Math.round((remaining / total) * 100);
-      const barLen = 20;
-      const filled = Math.round((remaining / total) * barLen);
-      const bar = `${c.green}${'█'.repeat(filled)}${c.dim}${'░'.repeat(barLen - filled)}${c.reset}`;
-      console.log(`${c.bold}Free tier:${c.reset}  ${remaining}/${total} calls remaining`);
-      console.log(`            ${bar} ${pct}%`);
-      if (remaining === 0) {
-        console.log(`${c.yellow}→ Next calls will use x402 payment ($0.001/call)${c.reset}`);
-      }
-    }
-  } else {
-    const err = await res.json() as any;
-    throw new Error(err.error?.message || 'Failed to get status');
-  }
-}
-
-async function cmdExport(opts: ParsedArgs) {
-  const params = new URLSearchParams();
-  if (opts.namespace) params.set('namespace', opts.namespace);
-  params.set('limit', opts.limit || '1000');
-  let offset = 0;
-  const allMemories: any[] = [];
-  const limit = parseInt(opts.limit || '1000');
-
-  // Paginate through all memories
-  while (true) {
-    params.set('offset', String(offset));
-    const result = await request('GET', `/v1/memories?${params}`) as any;
-    const memories = result.memories || result.data || [];
-    allMemories.push(...memories);
-    if (memories.length < limit) break;
-    offset += limit;
-    if (!outputQuiet) process.stderr.write(`${c.dim}Fetched ${allMemories.length} memories...${c.reset}\r`);
-  }
-
-  const exportData = {
-    version: 1,
-    exported_at: new Date().toISOString(),
-    count: allMemories.length,
-    memories: allMemories,
-  };
-
-  // Always output JSON for export (it's data)
-  outputWrite(JSON.stringify(exportData, null, 2));
-  if (!outputQuiet) {
-    outputError(`${c.green}✓${c.reset} Exported ${allMemories.length} memories`);
-  }
-}
-
-async function cmdImport(opts: ParsedArgs) {
-  let jsonText: string;
-
-  if (opts.file) {
-    const fs = await import('fs');
-    jsonText = fs.readFileSync(opts.file, 'utf-8');
-  } else {
-    const stdin = await readStdin();
-    if (!stdin) throw new Error('Provide --file <path> or pipe JSON via stdin');
-    jsonText = stdin;
-  }
-
-  const data = JSON.parse(jsonText);
-  const memories = data.memories || data;
-  if (!Array.isArray(memories)) throw new Error('Invalid format: expected { memories: [...] } or [...]');
-
-  // Concurrency control for parallel imports (default 5 to avoid rate limits)
-  const concurrency = opts.concurrency ? parseInt(opts.concurrency) : 5;
-  const batchSize = Math.min(concurrency, memories.length);
-  
-  let imported = 0;
-  let failed = 0;
-  
-  // Process in batches for concurrent import
-  for (let i = 0; i < memories.length; i += batchSize) {
-    const batch = memories.slice(i, i + batchSize);
-    
-    const results = await Promise.allSettled(
-      batch.map(async (mem) => {
-        const body: Record<string, any> = { content: mem.content };
-        if (mem.importance !== undefined) body.importance = mem.importance;
-        if (mem.metadata) body.metadata = mem.metadata;
-        if (mem.namespace || opts.namespace) body.namespace = mem.namespace || opts.namespace;
-        await request('POST', '/v1/store', body);
-        return true;
-      })
-    );
-    
-    for (const result of results) {
-      if (result.status === 'fulfilled') {
-        imported++;
-      } else {
-        failed++;
-        if (process.env.DEBUG) console.error(`Failed to import: ${result.reason}`);
-      }
-    }
-    
-    if (!outputQuiet) {
-      process.stderr.write(`\r  ${progressBar(imported, memories.length)}`);
-    }
-  }
-
-  if (!outputQuiet) process.stderr.write('\n');
-
-  if (outputJson) {
-    out({ imported, failed, total: memories.length });
-  } else {
-    success(`Imported ${imported}/${memories.length} memories${failed ? ` (${c.red}${failed} failed${c.reset})` : ''}`);
-  }
-}
-
-async function cmdStats(opts: ParsedArgs) {
-  // Gather stats from list endpoint
-  const params = new URLSearchParams();
-  if (opts.namespace) params.set('namespace', opts.namespace);
-  params.set('limit', '1');
-  
-  const result = await request('GET', `/v1/memories?${params}`) as any;
-  const total = result.total ?? '?';
-
-  // Get free tier status too
-  const walletAuth = await getWalletAuthHeader();
-  const statusRes = await fetch(`${API_URL}/v1/free-tier/status`, {
-    headers: { 'x-wallet-auth': walletAuth }
-  });
-
-  let tierData: any = {};
-  if (statusRes.ok) {
-    tierData = await statusRes.json();
-  }
-
-  if (outputJson) {
-    out({
-      total_memories: total,
-      api_url: API_URL,
-      wallet: tierData.wallet || getAccount().address,
-      free_tier_remaining: tierData.free_tier_remaining,
-      free_tier_total: tierData.free_tier_total,
-    });
-  } else {
-    console.log(`${c.bold}MemoClaw Stats${c.reset}`);
-    console.log(`${c.dim}${'─'.repeat(40)}${c.reset}`);
-    console.log(`Memories:        ${c.cyan}${total}${c.reset}`);
-    console.log(`API:             ${c.dim}${API_URL}${c.reset}`);
-    console.log(`Wallet:          ${c.dim}${tierData.wallet || getAccount().address}${c.reset}`);
-    if (tierData.free_tier_remaining !== undefined) {
-      console.log(`Free calls left: ${c.cyan}${tierData.free_tier_remaining}${c.reset}/${tierData.free_tier_total}`);
-    }
-    if (opts.namespace) {
-      console.log(`Namespace:       ${c.cyan}${opts.namespace}${c.reset}`);
-    }
-  }
-}
-
-async function cmdGraph(id: string, opts: ParsedArgs) {
-  // Fetch memory and its relations, display as ASCII graph
-  const result = await request('GET', `/v1/memories/${id}`) as any;
-  const mem = result.memory || result;
-  const relResult = await request('GET', `/v1/memories/${id}/relations`) as any;
-  const relations = relResult.relations || [];
-
-  if (outputJson) {
-    out({ memory: mem, relations });
-    return;
-  }
-
-  const label = (m: any) => {
-    const text = (m.content || '').slice(0, 40);
-    return text.length < (m.content || '').length ? text + '…' : text;
-  };
-
-  const shortId = (s: string) => s?.slice(0, 8) || '?';
-
-  console.log();
-  console.log(`  ${c.bold}${c.cyan}[${shortId(mem.id)}]${c.reset} ${label(mem)}`);
-
-  if (relations.length === 0) {
-    console.log(`  ${c.dim}  └── (no relations)${c.reset}`);
-  } else {
-    for (let i = 0; i < relations.length; i++) {
-      const r = relations[i];
-      const isLast = i === relations.length - 1;
-      const branch = isLast ? '└' : '├';
-      const typeColor = {
-        contradicts: c.red, supersedes: c.yellow, supports: c.green,
-        derived_from: c.magenta, related_to: c.blue,
-      }[r.relation_type] || c.dim;
-      console.log(`  ${c.dim}  ${branch}──${c.reset} ${typeColor}${r.relation_type}${c.reset} ${c.dim}→${c.reset} ${c.cyan}[${shortId(r.target_id)}]${c.reset}`);
-    }
-  }
-  console.log();
-}
-
-// ─── Namespace Command ──────────────────────────────────────────────────────
-
-async function cmdNamespace(subcmd: string, rest: string[], opts: ParsedArgs) {
-  if (subcmd === 'list' || !subcmd) {
-    // List all namespaces by querying with different namespace filters
-    // Since there's no direct namespace list API, we list memories and extract namespaces
-    const params = new URLSearchParams({ limit: '1000' });
-    const result = await request('GET', `/v1/memories?${params}`) as any;
-    const memories = result.memories || result.data || [];
-    
-    const nsSet = new Set<string>();
-    for (const mem of memories) {
-      if (mem.namespace) nsSet.add(mem.namespace);
-    }
-    const namespaces = Array.from(nsSet).sort();
-    
-    if (outputJson) {
-      out({ namespaces, count: namespaces.length });
-    } else if (namespaces.length === 0) {
-      console.log(`${c.dim}No namespaces found.${c.reset}`);
-    } else {
-      table(namespaces.map(ns => ({ namespace: ns })), [{ key: 'namespace', label: 'NAMESPACE', width: 30 }]);
-      console.log(`${c.dim}─ ${namespaces.length} namespace${namespaces.length !== 1 ? 's' : ''}${c.reset}`);
-    }
-  } else if (subcmd === 'stats') {
-    // Show memory counts per namespace
-    const params = new URLSearchParams({ limit: '1000' });
-    const result = await request('GET', `/v1/memories?${params}`) as any;
-    const memories = result.memories || result.data || [];
-    
-    const nsCounts: Record<string, number> = { '': 0 };
-    for (const mem of memories) {
-      const ns = mem.namespace || '';
-      nsCounts[ns] = (nsCounts[ns] || 0) + 1;
-    }
-    
-    const rows = Object.entries(nsCounts)
-      .sort((a, b) => b[1] - a[1])
-      .map(([ns, count]) => ({ namespace: ns || '(default)', count: String(count) }));
-    
-    if (outputJson) {
-      out({ namespaces: rows });
-    } else {
-      table(rows, [
-        { key: 'namespace', label: 'NAMESPACE', width: 30 },
-        { key: 'count', label: 'COUNT', width: 10 },
-      ]);
-    }
-  } else {
-    throw new Error('Usage: namespace [list|stats]');
-  }
-}
-
-async function cmdPurge(opts: ParsedArgs) {
-  // Support both --force and --yes flags
-  const confirmed = opts.force || opts.yes;
-  
-  if (!confirmed) {
-    // In non-interactive mode without --force, abort
-    if (!process.stdin.isTTY) {
-      throw new Error('Use --force or --yes to confirm purge in non-interactive mode');
-    }
-    const readline = await import('readline');
-    const rl = readline.createInterface({ input: process.stdin, output: process.stdout });
-    const answer = await new Promise<string>(r => rl.question(
-      `${c.red}⚠ Delete ALL memories${opts.namespace ? ` in namespace "${opts.namespace}"` : ''}? Type "yes" to confirm: ${c.reset}`,
-      r
-    ));
-    rl.close();
-    if (answer.trim().toLowerCase() !== 'yes') {
-      console.log(`${c.dim}Aborted.${c.reset}`);
-      return;
-    }
-  }
-
-  // Paginate and delete all with safety limit to prevent infinite loops
-  const params = new URLSearchParams({ limit: '100' });
-  if (opts.namespace) params.set('namespace', opts.namespace);
-  let deleted = 0;
-  let failedInRow = 0;
-  const MAX_CONSECUTIVE_FAILURES = 3;
-
-  while (true) {
-    params.set('offset', '0');
-    const result = await request('GET', `/v1/memories?${params}`) as any;
-    const memories = result.memories || result.data || [];
-    if (memories.length === 0) break;
-
-    let batchDeleted = 0;
-    for (const mem of memories) {
-      try {
-        await request('DELETE', `/v1/memories/${mem.id}`);
-        deleted++;
-        batchDeleted++;
-        failedInRow = 0;
-        if (!outputQuiet) process.stderr.write(`\r  ${progressBar(deleted, result.total || deleted)}`);
-      } catch (e: any) {
-        if (process.env.DEBUG) console.error(`\nFailed to delete ${mem.id}: ${e.message}`);
-      }
-    }
-
-    // If no deletes succeeded in this batch, we're stuck — bail out
-    if (batchDeleted === 0) {
-      failedInRow++;
-      if (failedInRow >= MAX_CONSECUTIVE_FAILURES) {
-        warn(`Aborting: ${MAX_CONSECUTIVE_FAILURES} consecutive batches failed to delete any memories`);
-        break;
-      }
-    }
-  }
-
-  if (!outputQuiet) process.stderr.write('\n');
-  if (outputJson) {
-    out({ deleted });
-  } else {
-    success(`Purged ${deleted} memories`);
-  }
-}
-
-async function cmdCount(opts: ParsedArgs) {
-  const params = new URLSearchParams({ limit: '1' });
-  if (opts.namespace) params.set('namespace', opts.namespace);
-  const result = await request('GET', `/v1/memories?${params}`) as any;
-  const total = result.total ?? '?';
-
-  if (outputJson) {
-    out({ count: total, namespace: opts.namespace || null });
-  } else {
-    console.log(String(total));
-  }
-}
-
-// ─── Init Command ────────────────────────────────────────────────────────────
-
-async function cmdInit(opts: ParsedArgs) {
-  const configPath = CONFIG_FILE_JSON;
-  
-  // Check if config already exists
-  if (fs.existsSync(configPath) && !opts.force) {
-    const existing = JSON.parse(fs.readFileSync(configPath, 'utf-8'));
-    console.error(`${c.yellow}⚠${c.reset} Config already exists at ${c.cyan}${configPath}${c.reset}`);
-    console.error(`  Wallet: ${c.dim}${existing.address || '(unknown)'}${c.reset}`);
-    console.error(`  Use ${c.bold}--force${c.reset} to overwrite.`);
-    process.exit(1);
-  }
-
-  // Generate new wallet using viem
-  const { generatePrivateKey } = await import('viem/accounts');
-  const privateKey = generatePrivateKey();
-  const newAccount = privateKeyToAccount(privateKey);
-
-  const apiUrl = opts.url || 'https://api.memoclaw.com';
-
-  // Ensure config directory exists
-  if (!fs.existsSync(CONFIG_DIR_PATH)) {
-    fs.mkdirSync(CONFIG_DIR_PATH, { recursive: true });
-  }
-
-  // Save config
-  const config = {
-    privateKey: privateKey,
-    address: newAccount.address,
-    url: apiUrl,
-  };
-  fs.writeFileSync(configPath, JSON.stringify(config, null, 2), { mode: 0o600 });
-
-  if (outputJson) {
-    out({ address: newAccount.address, url: apiUrl, configPath });
-  } else {
-    console.log();
-    console.log(`${c.green}✓${c.reset} ${c.bold}MemoClaw initialized!${c.reset}`);
-    console.log();
-    console.log(`  ${c.bold}Wallet:${c.reset}  ${c.cyan}${newAccount.address}${c.reset}`);
-    console.log(`  ${c.bold}API:${c.reset}     ${c.dim}${apiUrl}${c.reset}`);
-    console.log(`  ${c.bold}Config:${c.reset}  ${c.dim}${configPath}${c.reset}`);
-    console.log();
-    console.log(`  Your wallet is your identity. No signup needed.`);
-    console.log(`  You get ${c.bold}1000 free API calls${c.reset}, then x402 micropayments.`);
-    console.log();
-    console.log(`  ${c.dim}Try: memoclaw store "Hello, MemoClaw!"${c.reset}`);
-  }
-}
-
-// ─── Migrate Command ─────────────────────────────────────────────────────────
-
-async function cmdMigrate(targetPath: string, opts: ParsedArgs) {
-  if (!targetPath) {
-    throw new Error('Path required. Usage: memoclaw migrate <path-to-file-or-directory>');
-  }
-
-  const resolvedPath = path.resolve(targetPath);
-  
-  if (!fs.existsSync(resolvedPath)) {
-    throw new Error(`Path not found: ${resolvedPath}`);
-  }
-
-  // Collect all .md files
-  const mdFiles: { filepath: string; filename: string }[] = [];
-  const stat = fs.statSync(resolvedPath);
-  
-  const IGNORED_DIRS = new Set([
-    'node_modules', '.git', '.next', '.nuxt', 'dist', 'build', '.output',
-    '__pycache__', '.venv', 'venv', 'env', '.env', '.tox',
-    '.cache', '.tmp', 'tmp', 'coverage', '.nyc_output',
-    'vendor', 'target', '.gradle', '.mvn',
-    'skills', '.openclaw', '.clawd',
-  ]);
-
-  if (stat.isDirectory()) {
-    const walk = (dir: string) => {
-      const entries = fs.readdirSync(dir, { withFileTypes: true });
-      for (const entry of entries) {
-        const full = path.join(dir, entry.name);
-        if (entry.isDirectory()) {
-          if (!IGNORED_DIRS.has(entry.name)) {
-            walk(full);
-          }
-        } else if (entry.name.endsWith('.md')) {
-          mdFiles.push({ filepath: full, filename: path.relative(resolvedPath, full) });
-        }
-      }
-    };
-    walk(resolvedPath);
-  } else if (resolvedPath.endsWith('.md')) {
-    mdFiles.push({ filepath: resolvedPath, filename: path.basename(resolvedPath) });
-  } else {
-    throw new Error('Path must be a .md file or a directory containing .md files');
-  }
-
-  if (mdFiles.length === 0) {
-    throw new Error('No .md files found');
-  }
-
-  if (!outputQuiet) {
-    console.log(`${c.blue}ℹ${c.reset} Found ${c.bold}${mdFiles.length}${c.reset} markdown file${mdFiles.length !== 1 ? 's' : ''}`);
-  }
-
-  // Send in batches of up to 50 files
-  const BATCH_SIZE = 5;
-  let totalCreated = 0;
-  let totalDeduplicated = 0;
-  let totalErrors = 0;
-  let filesProcessed = 0;
-
-  for (let i = 0; i < mdFiles.length; i += BATCH_SIZE) {
-    const batch = mdFiles.slice(i, i + BATCH_SIZE);
-    const files = batch.map(f => ({
-      filename: f.filename,
-      content: fs.readFileSync(f.filepath, 'utf-8'),
-    }));
-
-    try {
-      const result = await request('POST', '/v1/migrate', { files }) as any;
-      totalCreated += result.memories_created || 0;
-      totalDeduplicated += result.memories_deduplicated || 0;
-      filesProcessed += result.files_processed || 0;
-      if (result.errors) totalErrors += result.errors.length;
-
-      if (!outputQuiet && !outputJson) {
-        process.stderr.write(`\r  ${progressBar(Math.min(i + BATCH_SIZE, mdFiles.length), mdFiles.length)}`);
-      }
-    } catch (e: any) {
-      totalErrors += batch.length;
-      if (!outputQuiet) {
-        console.error(`\n${c.red}Error:${c.reset} Batch failed: ${e.message}`);
-      }
-    }
-  }
-
-  if (!outputQuiet && !outputJson) process.stderr.write('\n');
-
-  if (outputJson) {
-    out({ files_found: mdFiles.length, files_processed: filesProcessed, memories_created: totalCreated, memories_deduplicated: totalDeduplicated, errors: totalErrors });
-  } else {
-    console.log();
-    success(`Migration complete!`);
-    console.log(`  Files processed:      ${c.cyan}${filesProcessed}${c.reset}`);
-    console.log(`  Memories created:     ${c.green}${totalCreated}${c.reset}`);
-    console.log(`  Deduplicated:         ${c.dim}${totalDeduplicated}${c.reset}`);
-    if (totalErrors > 0) {
-      console.log(`  Errors:               ${c.red}${totalErrors}${c.reset}`);
-    }
-  }
-}
-
-async function cmdCompletions(shell: string) {
-  const commands = ['init', 'migrate', 'store', 'recall', 'search', 'list', 'get', 'update', 'delete', 'ingest', 'extract',
-    'context', 'consolidate', 'relations', 'suggested', 'status', 'export', 'import', 'stats', 'browse',
-    'completions', 'config', 'graph', 'purge', 'count', 'namespace', 'help'];
-  
-  const globalFlags = ['--help', '--version', '--json', '--quiet', '--namespace', '--limit', '--offset',
-    '--tags', '--format', '--pretty', '--watch', '--raw', '--force', '--output', '--truncate',
-    '--no-truncate', '--columns', '--sort-by', '--reverse', '--wide', '--concurrency', '--yes',
-    '--timeout', '--field', '--dry-run'];
-  
-  if (shell === 'bash') {
-    console.log(`# Add to ~/.bashrc:
-# eval "$(memoclaw completions bash)"
-_memoclaw() {
-  local cur="\${COMP_WORDS[COMP_CWORD]}"
-  local prev="\${COMP_WORDS[COMP_CWORD-1]}"
-  local cmds="${commands.join(' ')}"
-  local flags="${globalFlags.join(' ')}"
-
-  if [ "\$COMP_CWORD" -eq 1 ]; then
-    COMPREPLY=( $(compgen -W "\$cmds" -- "\$cur") )
-  elif [[ "\$cur" == -* ]]; then
-    COMPREPLY=( $(compgen -W "\$flags" -- "\$cur") )
-  elif [[ "\$prev" == "--format" || "\$prev" == "-f" ]]; then
-    COMPREPLY=( $(compgen -W "json table csv yaml" -- "\$cur") )
-  elif [[ "\${COMP_WORDS[1]}" == "relations" && "\$COMP_CWORD" -eq 2 ]]; then
-    COMPREPLY=( $(compgen -W "list create delete" -- "\$cur") )
-  elif [[ "\${COMP_WORDS[1]}" == "config" && "\$COMP_CWORD" -eq 2 ]]; then
-    COMPREPLY=( $(compgen -W "show check init path" -- "\$cur") )
-  elif [[ "\${COMP_WORDS[1]}" == "namespace" && "\$COMP_CWORD" -eq 2 ]]; then
-    COMPREPLY=( $(compgen -W "list stats" -- "\$cur") )
-  elif [[ "\${COMP_WORDS[1]}" == "completions" && "\$COMP_CWORD" -eq 2 ]]; then
-    COMPREPLY=( $(compgen -W "bash zsh fish" -- "\$cur") )
-  fi
-}
-complete -F _memoclaw memoclaw`);
-  } else if (shell === 'zsh') {
-    console.log(`# Add to ~/.zshrc:
-# eval "$(memoclaw completions zsh)"
-_memoclaw() {
-  local -a commands=(${commands.map(c => `'${c}'`).join(' ')})
-  local -a flags=(${globalFlags.map(f => `'${f}'`).join(' ')})
-
-  if (( CURRENT == 2 )); then
-    _describe 'command' commands
-  else
-    case \${words[2]} in
-      relations)  _values 'subcommand' list create delete ;;
-      config)     _values 'subcommand' show check init path ;;
-      namespace)  _values 'subcommand' list stats ;;
-      completions) _values 'shell' bash zsh fish ;;
-      *)
-        if [[ "\$PREFIX" == -* ]]; then
-          _describe 'flag' flags
-        fi
-        ;;
-    esac
-  fi
-}
-compdef _memoclaw memoclaw`);
-  } else if (shell === 'fish') {
-    console.log(`# Add to ~/.config/fish/completions/memoclaw.fish:
-${commands.map(cmd => `complete -c memoclaw -n '__fish_use_subcommand' -a '${cmd}'`).join('\n')}
-
-# Subcommands
-complete -c memoclaw -n '__fish_seen_subcommand_from relations' -a 'list create delete'
-complete -c memoclaw -n '__fish_seen_subcommand_from config' -a 'show check init path'
-complete -c memoclaw -n '__fish_seen_subcommand_from namespace' -a 'list stats'
-complete -c memoclaw -n '__fish_seen_subcommand_from completions' -a 'bash zsh fish'
-
-# Global flags
-${globalFlags.map(f => `complete -c memoclaw -l '${f.replace(/^--/, '')}'`).join('\n')}`);
-  } else {
-    throw new Error(`Unknown shell: ${shell}. Supported: bash, zsh, fish`);
-  }
-}
-
-// ─── Config File Support ───────────────────────────────────────────────────
-
-interface ConfigFile {
-  url?: string;
-  privateKey?: string;
-  namespace?: string;
-  timeout?: number;
-}
-
-function loadConfigFile(): ConfigFile {
-  try {
-    if (fs.existsSync(CONFIG_FILE)) {
-      const content = fs.readFileSync(CONFIG_FILE, 'utf-8');
-      // Support both YAML and JSON
-      if (CONFIG_FILE.endsWith('.json')) {
-        return JSON.parse(content);
-      }
-      return yaml.load(content) as ConfigFile;
-    }
-  } catch (e: any) {
-    if (process.env.DEBUG) {
-      console.error(`Failed to load config: ${e.message}`);
-    }
-  }
-  return {};
-}
-
-function ensureConfigDir() {
-  if (!fs.existsSync(CONFIG_DIR)) {
-    fs.mkdirSync(CONFIG_DIR, { recursive: true });
-  }
-}
-
-// Load config file at startup
-const configFile = loadConfigFile();
-
-// ─── Config Command ─────────────────────────────────────────────────────────
-
-async function cmdConfig(subcmd: string, rest: string[]) {
-  if (subcmd === 'init') {
-    // Initialize config file
-    ensureConfigDir();
-    const sampleConfig: ConfigFile = {
-      url: 'https://api.memoclaw.com',
-      namespace: '',
-      timeout: 30,
-    };
-    
-    // If user has env vars set, use those as defaults
-    if (process.env.MEMOCLAW_URL) sampleConfig.url = process.env.MEMOCLAW_URL;
-    if (process.env.MEMOCLAW_NAMESPACE) sampleConfig.namespace = process.env.MEMOCLAW_NAMESPACE;
-    if (process.env.MEMOCLAW_TIMEOUT) sampleConfig.timeout = parseInt(process.env.MEMOCLAW_TIMEOUT);
-    
-    const configPath = CONFIG_FILE.endsWith('.yaml') || CONFIG_FILE.endsWith('.yml') 
-      ? CONFIG_FILE 
-      : CONFIG_FILE + '.yaml';
-    
-    fs.writeFileSync(configPath, yaml.dump(sampleConfig, { indent: 2 }));
-    success(`Config file created at ${c.cyan}${configPath}${c.reset}`);
-    console.log(`${c.dim}Edit this file and remove the privateKey line (set via MEMOCLAW_PRIVATE_KEY env var)${c.reset}`);
-    return;
-  }
-  
-  if (subcmd === 'path') {
-    console.log(CONFIG_FILE);
-    return;
-  }
-
-  if (subcmd === 'show' || !subcmd) {
-    const config: Record<string, string> = {
-      MEMOCLAW_URL: API_URL,
-      MEMOCLAW_PRIVATE_KEY: PRIVATE_KEY ? `${PRIVATE_KEY.slice(0, 6)}…${PRIVATE_KEY.slice(-4)}` : '(not set)',
-      NO_COLOR: process.env.NO_COLOR || '(not set)',
-      DEBUG: process.env.DEBUG || '(not set)',
-    };
-    if (outputJson) {
-      out(config);
-    } else {
-      console.log(`${c.bold}MemoClaw Configuration${c.reset}`);
-      console.log(`${c.dim}${'─'.repeat(50)}${c.reset}`);
-      for (const [key, val] of Object.entries(config)) {
-        const isSet = !val.includes('not set');
-        console.log(`  ${c.cyan}${key.padEnd(24)}${c.reset} ${isSet ? val : `${c.dim}${val}${c.reset}`}`);
-      }
-      console.log(`\n${c.dim}Set via environment variables or .env file${c.reset}`);
-    }
-  } else if (subcmd === 'check') {
-    // Validate configuration
-    const issues: string[] = [];
-    if (!PRIVATE_KEY) issues.push('MEMOCLAW_PRIVATE_KEY is not set');
-    else if (!PRIVATE_KEY.startsWith('0x')) issues.push('MEMOCLAW_PRIVATE_KEY should start with 0x');
-    else if (PRIVATE_KEY.length !== 66) issues.push(`MEMOCLAW_PRIVATE_KEY has wrong length (${PRIVATE_KEY.length}, expected 66)`);
-
-    if (outputJson) {
-      out({ valid: issues.length === 0, issues });
-    } else {
-      if (issues.length === 0) {
-        success('Configuration looks good!');
-        try {
-          const acct = getAccount();
-          info(`Wallet address: ${acct.address}`);
-        } catch {}
-      } else {
-        for (const issue of issues) {
-          console.log(`${c.red}✗${c.reset} ${issue}`);
-        }
-      }
-    }
-  } else {
-    throw new Error('Usage: config [show|check|init|path]');
-  }
-}
-
-async function cmdBrowse(opts: ParsedArgs) {
-  const readline = await import('readline');
-  const rl = readline.createInterface({ input: process.stdin, output: process.stdout });
-  const prompt = (q: string): Promise<string> => new Promise(r => rl.question(q, r));
-
-  console.log(`${c.bold}MemoClaw Interactive Browser${c.reset} ${c.dim}(type "help" or "q" to quit)${c.reset}`);
-  if (opts.namespace) console.log(`${c.dim}Namespace: ${opts.namespace}${c.reset}`);
-  console.log();
-
-  let offset = 0;
-  const limit = 10;
-
-  while (true) {
-    const input = (await prompt(`${c.cyan}memoclaw>${c.reset} `)).trim();
-    if (!input) continue;
-    if (input === 'q' || input === 'quit' || input === 'exit') break;
-
-    const parts = input.split(/\s+/);
-    const browsCmd = parts[0];
-    const browseArgs = parts.slice(1).join(' ');
-
-    try {
-      switch (browsCmd) {
-        case 'help':
-          console.log(`${c.bold}Commands:${c.reset}
-  list / ls          List memories (paginated)
-  next / n           Next page
-  prev / p           Previous page
-  get <id>           Show memory details
-  recall <query>     Search memories
-  store <content>    Store a new memory
-  delete <id>        Delete a memory
-  stats              Show stats
-  q / quit           Exit browser`);
-          break;
-        case 'list': case 'ls': {
-          offset = 0;
-          const params = new URLSearchParams({ limit: String(limit), offset: String(offset) });
-          if (opts.namespace) params.set('namespace', opts.namespace);
-          const result = await request('GET', `/v1/memories?${params}`) as any;
-          const memories = result.memories || result.data || [];
-          if (memories.length === 0) { console.log(`${c.dim}No memories.${c.reset}`); break; }
-          for (const m of memories) {
-            const text = m.content?.length > 60 ? m.content.slice(0, 60) + '…' : (m.content || '');
-            console.log(`  ${c.cyan}${(m.id || '?').slice(0, 8)}${c.reset}  ${text}`);
-          }
-          console.log(`${c.dim}─ showing ${offset + 1}-${offset + memories.length}${result.total ? ` of ${result.total}` : ''}${c.reset}`);
-          break;
-        }
-        case 'next': case 'n':
-          offset += limit;
-          /* falls through to list logic */ {
-            const params = new URLSearchParams({ limit: String(limit), offset: String(offset) });
-            if (opts.namespace) params.set('namespace', opts.namespace);
-            const result = await request('GET', `/v1/memories?${params}`) as any;
-            const memories = result.memories || result.data || [];
-            if (memories.length === 0) { console.log(`${c.dim}No more memories.${c.reset}`); offset = Math.max(0, offset - limit); break; }
-            for (const m of memories) {
-              const text = m.content?.length > 60 ? m.content.slice(0, 60) + '…' : (m.content || '');
-              console.log(`  ${c.cyan}${(m.id || '?').slice(0, 8)}${c.reset}  ${text}`);
-            }
-            console.log(`${c.dim}─ showing ${offset + 1}-${offset + memories.length}${result.total ? ` of ${result.total}` : ''}${c.reset}`);
-          }
-          break;
-        case 'prev': case 'p':
-          offset = Math.max(0, offset - limit);
-          {
-            const params = new URLSearchParams({ limit: String(limit), offset: String(offset) });
-            if (opts.namespace) params.set('namespace', opts.namespace);
-            const result = await request('GET', `/v1/memories?${params}`) as any;
-            const memories = result.memories || result.data || [];
-            for (const m of memories) {
-              const text = m.content?.length > 60 ? m.content.slice(0, 60) + '…' : (m.content || '');
-              console.log(`  ${c.cyan}${(m.id || '?').slice(0, 8)}${c.reset}  ${text}`);
-            }
-            console.log(`${c.dim}─ showing ${offset + 1}-${offset + memories.length}${result.total ? ` of ${result.total}` : ''}${c.reset}`);
-          }
-          break;
-        case 'get':
-          if (!browseArgs) { console.log(`${c.red}Usage: get <id>${c.reset}`); break; }
-          await cmdGet(browseArgs);
-          break;
-        case 'recall': case 'search':
-          if (!browseArgs) { console.log(`${c.red}Usage: recall <query>${c.reset}`); break; }
-          await cmdRecall(browseArgs, opts);
-          break;
-        case 'store':
-          if (!browseArgs) { console.log(`${c.red}Usage: store <content>${c.reset}`); break; }
-          await cmdStore(browseArgs, opts);
-          break;
-        case 'delete': case 'rm':
-          if (!browseArgs) { console.log(`${c.red}Usage: delete <id>${c.reset}`); break; }
-          await cmdDelete(browseArgs);
-          break;
-        case 'stats':
-          await cmdStats(opts);
-          break;
-        default:
-          console.log(`${c.dim}Unknown command. Type "help" for available commands.${c.reset}`);
-      }
-    } catch (e: any) {
-      console.log(`${c.red}Error:${c.reset} ${e.message}`);
-    }
-    console.log();
-  }
-
-  rl.close();
-  console.log(`${c.dim}Bye!${c.reset}`);
-}
-
-// ─── Help ────────────────────────────────────────────────────────────────────
-
-function printHelp(command?: string) {
-  if (command) {
-    const subHelp: Record<string, string> = {
-      store: `${c.bold}memoclaw store${c.reset} "content" [options]
-
-Store a memory. Content can be provided as a positional argument,
-via --content flag, or piped via stdin.
-
-  ${c.dim}memoclaw store "Hello world"${c.reset}
-  ${c.dim}memoclaw store --content "Hello world"${c.reset}
-  ${c.dim}echo "Hello world" | memoclaw store${c.reset}
-
-Options:
-  --content <text>       Memory content (alternative to positional arg)
-  --importance <0-1>     Importance score (default: 0.5)
-  --tags <tag1,tag2>     Comma-separated tags
-  --namespace <name>     Memory namespace`,
-
-      search: `${c.bold}memoclaw search${c.reset} "query" [options]
-
-Free text search (no embeddings cost). Use this for exact/keyword matching.
-For semantic similarity search, use \`recall\` instead.
-
-Options:
-  --limit <n>            Max results (default: 10)
-  --namespace <name>     Filter by namespace
-  --tags <tag1,tag2>     Filter by tags
-  --raw                  Output content only (for piping)`,
-
-      context: `${c.bold}memoclaw context${c.reset} "query" [options]
-
-Get a GPT-powered contextual summary from your memories.
-Uses GPT-4o-mini + embeddings. Costs $0.01 per call.
-
-  ${c.dim}memoclaw context "What do I know about project X?"${c.reset}
-
-Options:
-  --namespace <name>     Filter by namespace
-  --limit <n>            Max memories to consider`,
-
-      recall: `${c.bold}memoclaw recall${c.reset} "query" [options]
-
-Search memories by semantic similarity.
-
-Options:
-  --limit <n>            Max results (default: 10)
-  --min-similarity <0-1> Similarity threshold (default: 0.5)
-  --namespace <name>     Filter by namespace
-  --tags <tag1,tag2>     Filter by tags`,
-
-      list: `${c.bold}memoclaw list${c.reset} [options]
-
-List all memories in a table format.
-
-Options:
-  --limit <n>            Max results (default: 20)
-  --offset <n>           Pagination offset
-  --namespace <name>     Filter by namespace
-  --sort-by <field>      Sort by field (id, importance, created, updated)
-  --reverse              Reverse sort order
-  --columns <cols>       Select columns (id,content,importance,tags,created)`,
-
-      export: `${c.bold}memoclaw export${c.reset} [options]
-
-Export all memories as JSON. Useful for backups.
-
-  ${c.dim}memoclaw export > backup.json${c.reset}
-  ${c.dim}memoclaw export --namespace project1 > project1.json${c.reset}
-
-Options:
-  --namespace <name>     Filter by namespace
-  --limit <n>            Max per page (default: 1000)`,
-
-      import: `${c.bold}memoclaw import${c.reset} [options]
-
-Import memories from JSON file or stdin.
-
-  ${c.dim}memoclaw import --file backup.json${c.reset}
-  ${c.dim}cat backup.json | memoclaw import${c.reset}
-
-Options:
-  --file <path>          JSON file to import
-  --namespace <name>     Override namespace for all memories`,
-
-      stats: `${c.bold}memoclaw stats${c.reset} [options]
-
-Show memory statistics and account info.
-
-Options:
-  --namespace <name>     Filter by namespace`,
-
-      get: `${c.bold}memoclaw get${c.reset} <id>
-
-Retrieve a single memory by its ID.`,
-
-      config: `${c.bold}memoclaw config${c.reset} [show|check]
-
-Show or validate your MemoClaw configuration.
-
-Subcommands:
-  show       Display current configuration (default)
-  check      Validate configuration and test connectivity`,
-
-      browse: `${c.bold}memoclaw browse${c.reset} [options]
-
-Interactive memory browser (REPL). Explore, search, and manage
-memories in a persistent session.
-
-Options:
-  --namespace <name>     Filter by namespace
-
-Commands inside browser: list, get, recall, store, delete, stats, next, prev`,
-
-      graph: `${c.bold}memoclaw graph${c.reset} <id>
-
-Show an ASCII tree of a memory and its relations.
-
-Options:
-  --json                 Output as JSON`,
-
-      purge: `${c.bold}memoclaw purge${c.reset} [options]
-
-Delete ALL memories. Requires confirmation or --force.
-
-  ${c.dim}memoclaw purge --force${c.reset}
-  ${c.dim}memoclaw purge --namespace old-project --force${c.reset}
-
-Options:
-  --force                Skip confirmation prompt
-  --namespace <name>     Only purge memories in namespace`,
-
-      count: `${c.bold}memoclaw count${c.reset} [options]
-
-Print the total number of memories (pipe-friendly).
-
-  ${c.dim}memoclaw count${c.reset}
-  ${c.dim}memoclaw count --namespace project1${c.reset}
-
-Options:
-  --namespace <name>     Count only in namespace`,
-
-      completions: `${c.bold}memoclaw completions${c.reset} <bash|zsh|fish>
-
-Generate shell completion scripts.
-
-  ${c.dim}eval "$(memoclaw completions bash)"${c.reset}
-  ${c.dim}eval "$(memoclaw completions zsh)"${c.reset}
-  ${c.dim}memoclaw completions fish > ~/.config/fish/completions/memoclaw.fish${c.reset}`,
-
-      namespace: `${c.bold}memoclaw namespace${c.reset} [list|stats]
-
-Manage and view namespaces.
-
-Subcommands:
-  list       List all unique namespaces (default)
-  stats      Show memory counts per namespace
-
-  ${c.dim}memoclaw namespace list${c.reset}
-  ${c.dim}memoclaw namespace stats${c.reset}`,
-    };
-
-    if (subHelp[command]) {
-      console.log(subHelp[command]);
-    } else {
-      console.log(`No detailed help for "${command}". Run ${c.dim}memoclaw --help${c.reset} for overview.`);
-    }
-    return;
-  }
-
-  console.log(`${c.bold}MemoClaw CLI${c.reset} ${c.dim}v${VERSION}${c.reset} — Memory-as-a-Service for AI agents
-
-${c.bold}Usage:${c.reset}
-  memoclaw <command> [options]
-
-${c.bold}Commands:${c.reset}
-  ${c.cyan}init${c.reset}                   Generate wallet & initialize config
-  ${c.cyan}migrate${c.reset} <path>          Import .md files from OpenClaw/local
-  ${c.cyan}store${c.reset} "content"        Store a memory (also accepts stdin)
-  ${c.cyan}recall${c.reset} "query"         Search memories by similarity
-  ${c.cyan}search${c.reset} "query"         Text search (free, no embeddings cost)
-  ${c.cyan}list${c.reset}                   List memories in a table
-  ${c.cyan}get${c.reset} <id>               Get a single memory by ID
-  ${c.cyan}update${c.reset} <id>            Update a memory
-  ${c.cyan}delete${c.reset} <id>            Delete a memory
-  ${c.cyan}ingest${c.reset}                 Ingest raw text into memories
-  ${c.cyan}extract${c.reset} "text"         Extract memories from text
-  ${c.cyan}context${c.reset} "query"        Get GPT-powered contextual summary ($0.01/call)
-  ${c.cyan}consolidate${c.reset}            Merge similar memories
-  ${c.cyan}relations${c.reset} <sub>        Manage memory relations
-  ${c.cyan}suggested${c.reset}              Get suggested memories for review
-  ${c.cyan}status${c.reset}                 Check account & free tier info
-  ${c.cyan}stats${c.reset}                  Memory statistics
-  ${c.cyan}export${c.reset}                 Export all memories as JSON
-  ${c.cyan}import${c.reset}                 Import memories from JSON
-  ${c.cyan}completions${c.reset} <shell>    Generate shell completions
-  ${c.cyan}browse${c.reset}                 Interactive memory browser (REPL)
-  ${c.cyan}config${c.reset} [show|check]    Show or validate configuration
-  ${c.cyan}graph${c.reset} <id>             ASCII visualization of memory relations
-  ${c.cyan}purge${c.reset}                  Delete ALL memories (requires --force or confirm)
-  ${c.cyan}namespace${c.reset} [list|stats] Manage and view namespaces
-  ${c.cyan}count${c.reset}                  Quick memory count
-  ${c.cyan}help${c.reset} [command]          Show help for a command
-
-${c.bold}Global Options:${c.reset}
-  -h, --help             Show help (use with command for details)
-  -v, --version          Show version
-  -j, --json             Output as JSON (machine-readable)
-  -q, --quiet            Suppress non-essential output
-  -n, --namespace <name> Filter/set namespace
-  -l, --limit <n>        Limit results
-  -o, --offset <n>      Pagination offset
-  -t, --tags <a,b>       Comma-separated tags
-  -f, --format <fmt>     Output format: json, table, csv, yaml
-  -p, --pretty          Pretty-print JSON output
-  -w, --watch            Watch for changes (continuous polling)
-  --watch-interval <ms>  Polling interval for watch mode (default: 5000)
-  -s, --truncate <n>     Truncate output to n characters
-  --no-truncate         Disable truncation in output
-  -c, --concurrency <n>  Number of parallel imports (default: 1)
-  -y, --yes              Skip confirmation prompts (alias for --force)
-  -O, --output <file>    Write output to file instead of stdout
-  -F, --field <name>     Extract specific field from output
-  -r, --reverse          Reverse sort order
-  -m, --sort-by <field>  Sort by field (id, importance, created, updated)
-  -k, --columns <cols>   Select columns (id,content,importance,tags,created)
-  --raw                  Raw output (content only, for piping)
-  --wide                 Use wider columns in table output
-  --force                Skip confirmation prompts
-  -T, --timeout <sec>    Request timeout (default: 30)
-
-${c.bold}Environment:${c.reset}
-  MEMOCLAW_PRIVATE_KEY   Wallet private key for auth + payments
-  MEMOCLAW_URL           API endpoint (default: https://api.memoclaw.com)
-  NO_COLOR               Disable colored output
-  DEBUG                  Enable debug logging
-
-${c.bold}Piping:${c.reset}
-  echo "meeting notes" | memoclaw store
-  echo "long text" | memoclaw ingest
-  memoclaw export | jq '.memories | length'
-  cat backup.json | memoclaw import
-
-${c.bold}Free Tier:${c.reset}
-  Every wallet gets 1000 free API calls. After that, x402
-  micropayments kick in automatically ($0.001/call USDC on Base).
-
-${c.dim}API: https://api.memoclaw.com${c.reset}`);
-}
+import { VERSION } from './config.js';
+import { c } from './colors.js';
+import { configureOutput, outputJson, readStdin } from './output.js';
+import { setRequestTimeout } from './http.js';
+import { printHelp } from './help.js';
+
+// Commands
+import { cmdStore } from './commands/store.js';
+import { cmdRecall } from './commands/recall.js';
+import { cmdList } from './commands/list.js';
+import { cmdGet, cmdDelete, cmdUpdate } from './commands/memory.js';
+import { cmdSearch, cmdContext, cmdExtract, cmdIngest, cmdConsolidate } from './commands/search.js';
+import { cmdRelations } from './commands/relations.js';
+import { cmdStatus, cmdStats, cmdCount, cmdSuggested, cmdGraph } from './commands/status.js';
+import { cmdExport, cmdImport, cmdPurge } from './commands/data.js';
+import { cmdNamespace } from './commands/namespace.js';
+import { cmdInit, cmdConfig } from './commands/config.js';
+import { cmdMigrate } from './commands/migrate.js';
+import { cmdBrowse } from './commands/browse.js';
+import { cmdCompletions } from './commands/completions.js';
 
 // ─── Main ────────────────────────────────────────────────────────────────────
 
 const args = parseArgs(process.argv.slice(2));
 const [cmd, ...rest] = args._;
 
-outputJson = !!args.json;
-outputQuiet = !!args.quiet;
-outputPretty = !!args.pretty;
-
-// Parse field extraction (--field implies --json to ensure structured output)
-if (args.field && args.field !== true) {
-  outputField = String(args.field);
-  outputJson = true;
-}
-
-// Parse output format
-if (args.format) {
-  let fmt = String(args.format).toLowerCase();
-  // Normalize yml to yaml
-  if (fmt === 'yml') fmt = 'yaml';
-  if (fmt === 'json' || fmt === 'table' || fmt === 'csv' || fmt === 'yaml') {
-    outputFormat = fmt;
-  }
-}
-// --json flag overrides format
-if (args.json) {
-  outputFormat = 'json';
-}
-
-// Parse truncate option
-if (args.truncate != null && args.truncate !== true) {
-  outputTruncate = parseInt(args.truncate);
-} else if (args.truncate === true) {
-  outputTruncate = 80; // default truncate width
-}
-
-// Parse no-truncate option (disables truncation)
-noTruncate = !!args.noTruncate;
-if (noTruncate) {
-  outputTruncate = 0;
-}
-
-// Parse output file
-if (args.output) {
-  outputFile = String(args.output);
-  // Clear the file first
-  fs.writeFileSync(outputFile, '');
-}
+// Configure output state
+configureOutput(args);
 
 if (args.version) {
   console.log(`memoclaw ${VERSION}`);
@@ -1977,8 +56,9 @@ if (args.help) {
   process.exit(0);
 }
 
-// Wrap request() with timeout support
+// Configure request timeout
 const TIMEOUT_MS = args.timeout ? parseInt(args.timeout) * 1000 : 30000;
+setRequestTimeout(TIMEOUT_MS);
 
 try {
   switch (cmd) {
