+++ src/colors.ts
@@ -1,0 +1,18 @@
+/**
+ * Terminal color utilities
+ */
+
+const NO_COLOR = !!process.env.NO_COLOR || !process.stdout.isTTY;
+
+export const c = {
+  reset: NO_COLOR ? '' : '\x1b[0m',
+  bold: NO_COLOR ? '' : '\x1b[1m',
+  dim: NO_COLOR ? '' : '\x1b[2m',
+  red: NO_COLOR ? '' : '\x1b[31m',
+  green: NO_COLOR ? '' : '\x1b[32m',
+  yellow: NO_COLOR ? '' : '\x1b[33m',
+  blue: NO_COLOR ? '' : '\x1b[34m',
+  magenta: NO_COLOR ? '' : '\x1b[35m',
+  cyan: NO_COLOR ? '' : '\x1b[36m',
+  gray: NO_COLOR ? '' : '\x1b[90m',
+};
